### 2017-03-04 #192 @whedon
[PRE REVIEW]: Fitting a Gamma-Gompertz survival model to capture-recapture data collected on free-ranging animal populations
**Submitting author:** oliviergimenez (<a href="http://orcid.org/0000-0001-7001-5142">Olivier Gimenez</a>)
**Repository:** <a href="https://github.com/oliviergimenez/GammaGompertzCR" target ="_blank">https://github.com/oliviergimenez/GammaGompertzCR</a>
**Version:** v1.0.0

Thanks for submitting your paper to JOSS oliviergimenez. The JOSS editor (shown at the top of this issue) will work with you on this issue to find a reviewer for your submission before creating the main review issue.
