### 2017-03-20 #216 @whedon
[REVIEW]: Fitting a Gamma-Gompertz survival model to capture-recapture data collected on free-ranging animal populations
**Submitting author:** @oliviergimenez (<a href="http://orcid.org/0000-0001-7001-5142">Olivier Gimenez</a>)
**Repository:** <a href="https://github.com/oliviergimenez/GammaGompertzCR" target ="_blank">https://github.com/oliviergimenez/GammaGompertzCR</a>
**Version:** v1.0.0

- [x] **Authorship:** Has the submitting author (oliviergimenez) made major contributions to the software?
