### 2017-04-17 #216 @bbolker
it had slipped but am working on it and think I should be able to get it done shortly.  How granular would you like the issues in the target-repo issue list to be?  (One long issue with lots of bullet points?  One issue per item above? Somewhere in between?)
