### 2017-04-18 oliviergimenez/GammaGompertzCR#2 @bbolker
review for JOSS (1/5)
Overall, I think this package is fine (clear and, as far as I can see, technically correct).  There are a handful of changes that need to be made to conform to JOSS's standards, but those should be relatively easy. On the other hand, there are a rather large number of chores, individually not too big deal but overall perhaps rather overwhelming, that could be made to make this package *much* easier to use and enhance its uptake in the research community.
