### 2017-04-18 oliviergimenez/GammaGompertzCR#3 @bbolker
review for JOSS (2/5): conforming to JOSS guidelines (license, community guidelines)
There are a few particular items that seem to need to be done to conform to JOSS's standards. Some are a little odd when looked at from the R-package point of view, but I would fix them rather than make an exception for R packages.

- the package does not "contain a plain-text LICENSE file with the contents of an OSI approved software license".  The DESCRIPTION file does indicate "MIT", which is the standard way to specify licensing in an R package (but note that `R CMD check` does note: "License components which are templates and need '+ file LICENSE': MIT"; see https://cran.r-project.org/web/licenses/MIT
- **Community guidelines**: "Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support".  No. It would be a reasonable guess that (1) forking and submitting pull requests; (2) posting to the Issues list on GitHub would be reasonable approaches, but this could be made explicit. (The `DESCRIPTION` file in an R package also allows for a `URL:` field and a `BugReports:` field.)
