### 2017-04-18 oliviergimenez/GammaGompertzCR#4 @bbolker
review for JOSS (3/5): tests
The package contains no tests.  This could be done by the standard old-school R approach (`.R` files in `tests/`), or via the `testthat` unit-testing framework.

Some thought will have to be given to the degree of reproducibility required (tolerances, random-number seed setting for JAGS, etc.) and to how to make the tests run fast enough (e.g. run for minimal numbers of steps).

The manual example in the `README` file does serve as a minimal example to make sure the software runs, but it does not show output (nor set random-number seeds) so there is no way to know if the output obtained is correct.

There are also no examples given in the manual.  The examples given in the `README` file should probably be incorporated in the package help pages.
