### 2017-04-18 oliviergimenez/GammaGompertzCR#5 @bbolker
review for JOSS (4/5): documentation of functionality
The documentation of the functionality in the package is relatively sparse. The minimal pieces are there (README and help pages for the three functions contained in the package), and there are pointers to the literature so that users can in principle go educate themselves on the technical details underpinning the computations.  However, what's here is extremely bare-bones, and it would be kind to users to give them a bit more help:

- describe the level of expertise in Bayesian MCMC techniques expected as a prerequisite for using the package safely (and possibly pointers to ecologist-friendly introductions by Kéry, McCarthy, ...). For example, users need to know how to interpret standard MCMC diagnostics (trace plots, Gelman-Rubin statistics, etc.) and what to do when they fail.
- Are there any diagnostics (e.g., sensible ranges of parameters) that are specific to this model, as opposed to the general tools that apply to most MCMC approaches?)
- What priors are used? Are the results sensitive to the priors? Can users override the default priors? When might they want to?
- How are initial values chosen for the chains? Can users override the defaults? When might they want to?
- Are there reasonable minimum and maximum data set sizes for this approach?
- Is there any capability for including fixed-effect covariates or random-effects groupings in the model? (I realize this may be beyond the scope of this package, but it might be worth mentioning)
