### 2017-04-18 oliviergimenez/GammaGompertzCR#6 @bbolker
review for JOSS (5/5): bits and pieces
I don't consider any of these issues publication-blockers, but many of them are fairly low-hanging fruit.
- in its current form the package does not pass `R CMD check ...` (at least with a development version of R) - and so it cannot be put on CRAN
- there are many accessor methods that could be provided to make users' lives easier.
     - `print()`  (to print short summary of model)
     - `coef(summary(...))` (to return coefficient table)
     - `plot()` (to plot model diagnostics?)  (I have a personal preference for `coda::xyplot.mcmc` (trace plots only, compact/lattice format) and `coda::densityplot.mcmc` (density plots, only, compact/lattice format) over `plot.mcmc`  (trace & density plots)
     - in the README, why use `post_inf[[1]]` (first chain only) rather than output from `dclone::coef()` to get parameter values
     - storing output from longer runs (i.e., with larger numbers of clones) in `inst` would make it easier to give examples that illustrate convergence of data-cloning in a non-trivial way
     - why not include the dipper data as a a data object in the package?
     - the examples print a (skipped) error "Error in chol.default(W) : the leading minor of order 6 is not positive definite". This is probably trivial, but might worry users.
     - FWIW my fork of the `broom` package (https://github.com/bbolker/broom) allows "tidying" of `mcmc` objects, so e.g. `mm <- emdbook::lump.mcmc.list(post_inf); broom::tidy(mm); dotwhisker::dwplot(mm)` produces useful output
