### 2017-04-18 #216 @bbolker
I'm not sure if this is how I'm supposed to do it, but these issues constitute my "review"

- https://github.com/oliviergimenez/GammaGompertzCR/issues/2
- https://github.com/oliviergimenez/GammaGompertzCR/issues/3
- https://github.com/oliviergimenez/GammaGompertzCR/issues/4
- https://github.com/oliviergimenez/GammaGompertzCR/issues/5 (not necessarily blocking)
- https://github.com/oliviergimenez/GammaGompertzCR/issues/6 (non-blocking)
