### 2017-04-20 #216 @karthik
@bbolker
🙏 a ton!

We don't have a set review style and we leave it up to the reviewers.

@oliviergimenez Can you please address the issues raised by Dr. Bolker in those individual threads and present us with a summary here once you are done. That way @bbolker can check that he is satisfied with the changes and recommend an acceptance.
