### 2017-04-20 #216 @karthik
@bbolker PS: I'm sure you must have seen this GitHub trick, but if you mention `openjournals/joss-reviews/#216` anywhere is an issue on the author's repo, it automatically lists them here with you having to manually link them in the body of your review. That trick also works across repos. Syntax is just `user/repo/#issue_number`
