### 2017-04-20 #216 @oliviergimenez
Thanks a lot @bbolker. My co-author and I will address the issues you raised the best we can. I'm going to be busy over the next two weeks though, is the end of May working for you @karthik ? Cheers, Olivier.
