+++ README.md
@@ -1,8 +1,10 @@
 ---
 title: "Fitting a Gamma-Gompertz survival model to capture-recapture data collected on free-ranging animal populations"
 author: "Gilbert Marzolin and Olivier Gimenez"
-date: "25 février 2017"
-output: html_document
+date: "12 January 2018"
+output: 
+  html_document: 
+    keep_md: yes
 ---
 
 # In brief
@@ -10,149 +12,256 @@ output: html_document
 `GammaGompertzCR` is an R package that allows estimating survival in free-ranging animal populations using a Gompertz capture-recapture model with a Gamma frailty to deal with individual heterogeneity. 
 It uses data cloning in the Bayesian framework to get maximum likelihood parameter estimates (Lele et al. 2007). Data cloning uses multiple copies of the data to produce prior-invariant inferences and a normal distribution centered at the maximum likelihood estimates. In addition, this method allows detecting non-identifiable parameters (Lele et al. 2010).
 
-This repository hosts the development version of the package. Assuming JAGS is already on your system (otherwise: <http://mcmc-jags.sourceforge.net/>), you can install `GammaGompertzCR` as follows:
-```{r}
+# Installation
+
+This repository hosts the development version of the package. Assuming `JAGS` is already on your system (otherwise it'll take a minute to [get it](http://mcmc-jags.sourceforge.net/), you can install `GammaGompertzCR` as follows:
+
+```r
 if(!require(devtools)) install.packages("devtools")
 library("devtools")
 install_github('oliviergimenez/GammaGompertzCR')
 ```
 
+# Getting help, reporting an issue or contribute
+
+To report bugs/issues/feature requests, please file an [issue](https://github.com/oliviergimenez/GammaGompertzCR/issues). These are very welcome!
+
+Feel free to contribute to the repository by forking and submitting a pull request. If you are new to GitHub, you might start with a [basic tutorial](https://help.github.com/articles/set-up-git) and check out [a more detailed guide to pull requests](https://help.github.com/articles/about-pull-requests/).
+
+If you prefer to email, feel free to drop a message to the Gilbert and Olivier (see the DESCRIPTION file of this repository for their contact details).
+
 # A session example
 
+## Preliminary steps
+
 First, we need to load the package:
-```{r}
+
+```r
 library(GammaGompertzCR)
 ```
 
-Then, we read in some data. Below we use a fragment of the dipper dataset used by Marzolin et al. (2011). Detections are the 1's, non-detections the 2's.
-```{r}
-mydat <- matrix(c(1,1,1,1,1,1,1,1,2,
-1,1,1,1,1,1,1,2,2,
-1,1,1,1,1,1,2,2,2,
-1,1,1,1,1,2,2,2,2,
-1,1,1,2,2,2,2,2,2,
-1,1,1,2,2,2,2,2,2,
-1,1,1,2,2,2,2,2,2,
-1,1,2,1,2,2,2,2,2,
-1,1,2,2,2,2,2,2,2,
-1,1,2,2,2,2,2,2,2,
-1,1,2,2,2,2,2,2,2,
-1,2,1,1,1,1,1,2,2,
-1,2,1,1,1,2,2,2,2,
-1,2,1,1,2,1,1,2,2,
-1,2,1,1,2,2,2,2,2,
-1,2,1,2,2,2,2,2,2,
[…]
