+++ R/data.R
@@ -1,0 +1,15 @@
+#' Dipper dataset
+#'
+#' A dataset containing detections (1) and non-detections (2) of birds
+#'
+#' @format A matrix with 1176 rows (individuals) and 35 column (capture occasions):
+#' @source \url{http://www.esapubs.org/archive/ecol/E092/048/}
+"dipper_complete"
+
+#' A fragment of the Dipper dataset
+#'
+#' A dataset containing detections (1) and non-detections (2) of birds
+#'
+#' @format A matrix with 75 rows (individuals) and 9 column (capture occasions):
+#' @source \url{http://www.esapubs.org/archive/ecol/E092/048/}
+"dipper_tiny"
