+++ R/ggcr_model.R
@@ -1,7 +1,8 @@
-#' ggcr_model: Gamma-Gompertz capture-recapture model
+#' ggcr_model: Gamma-Gompertz capture-recapture model 
 #'
-#' This function implements a gamma-gompertz capture-recapture model in the BUGS syntax
+#' This function implements a gamma-gompertz capture-recapture model in the BUGS syntax. The survival of unit 'i' at age 'j' is expressed through the parameters of a Gompertz baseline mortality and a Gamma frailty. Then it is used along with the detection probability to tie the animal real states (alive or dead when unobserved) to the observed capture-recapture matrix, hence to deduce the survival parameters by maximum likelihood through data-cloning. The model is parameterized with a (coefficient of the baseline mortality rate), b (coefficient in the exponent of the baseline mortality rate) and k (inverse of variance of the Gamma function) for the Gamma-Gompertz survival along with the nuisance detection parameters psi (mean of logit detection) and sigeta (standard deviation of logit detection). Through marginalization (Missov 2013) the population survival function enables the plot of survival by ages. 
 #' @author Gilbert Marzolin, Olivier Gimenez
+#' @references Missov, TI. 2013. Gamma-Gompertz Life Expectancy at Birth. Demographic Research 28:259-270. doi:10.4054/DemRes.2013.28.9
 #' @keywords package
 #' @export
 
@@ -15,7 +16,7 @@ psi ~ dnorm(pri[4,1],pri[4,2])
 log.sigeta ~ dnorm(pri[5,1],pri[5,2])
 
 for (i in 1:nind){  # for each unit
-	st[i,first[i]] <- mydat[i,first[i]]
+	st[i,first[i]] <- mydata[i,first[i]]
 
 	for (j in first[i]:(nyears-1)){
 		
@@ -28,7 +29,7 @@ for (i in 1:nind){  # for each unit
         
         ## OBSERVATION EQUATIONS ##
 		# draw Obs[i,j+1] given st[i,j+1]
-		mydat[i,j+1]~ dcat(c(p[i,j],1-p[i,j])*equals(st[i,j+1],1)+ c(0,1)*equals(st[i,j+1],2))
+		mydata[i,j+1]~ dcat(c(p[i,j],1-p[i,j])*equals(st[i,j+1],1)+ c(0,1)*equals(st[i,j+1],2))
 		
 		# p[i,j] is detection prob at j+1
 		logit(p[i,j]) <- psi + eta[j]  #same detection for all units with time random effect.
