+++ man/dipper_tiny.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{dipper_tiny}
+\alias{dipper_tiny}
+\title{A fragment of the Dipper dataset}
+\format{A matrix with 75 rows (individuals) and 9 column (capture occasions):}
+\source{
+\url{http://www.esapubs.org/archive/ecol/E092/048/}
+}
+\usage{
+dipper_tiny
+}
+\description{
+A dataset containing detections (1) and non-detections (2) of birds
+}
+\keyword{datasets}
