+++ man/dipper_complete.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{dipper_complete}
+\alias{dipper_complete}
+\title{Dipper dataset}
+\format{A matrix with 1176 rows (individuals) and 35 column (capture occasions):}
+\source{
+\url{http://www.esapubs.org/archive/ecol/E092/048/}
+}
+\usage{
+dipper_complete
+}
+\description{
+A dataset containing detections (1) and non-detections (2) of birds
+}
+\keyword{datasets}
