+++ README.md
@@ -9,8 +9,7 @@ output:
 
 # In brief
   
-`GammaGompertzCR` is an R package that allows estimating survival in free-ranging animal populations using a Gompertz capture-recapture model with a Gamma frailty to deal with individual heterogeneity. 
-It uses data cloning in the Bayesian framework to get maximum likelihood parameter estimates (Lele et al. 2007). Data cloning uses multiple copies of the data to produce prior-invariant inferences and a normal distribution centered at the maximum likelihood estimates. In addition, this method allows detecting non-identifiable parameters (Lele et al. 2010).
+`GammaGompertzCR` is an R package that allows estimating survival in free-ranging animal populations using a Gompertz capture-recapture model with a Gamma frailty to deal with individual heterogeneity. It uses data cloning in the Bayesian framework to get maximum likelihood parameter estimates (Lele et al. 2007). Data cloning uses multiple copies of the data to produce prior-invariant inferences and a normal distribution centered at the maximum likelihood estimates. In addition, this method allows detecting non-identifiable parameters (Lele et al. 2010). To use the package, users should be familiar with Bayesian MCMC techniques and in particular how to interpret convergence diagnostics. We refer to Robert and Casella (2010) for an introduction to Bayesian MCMC techniques with R and to King et al. (2009) for an introduction for ecologists.
 
 # Installation
 
@@ -106,20 +105,20 @@ summary(post_inf[[length(clo)]])
 ##    plus standard error of the mean:
 ## 
 ##          Mean      SD  DC SD  Naive SE Time-series SE  R hat
-## a      0.1180 0.01696 0.1696 0.0003096      0.0003163 0.9994
-## b      0.4874 0.02824 0.2824 0.0005155      0.0005375 1.0011
-## k      4.8995 0.03127 0.3127 0.0005709      0.0005718 1.0002
-## psi    1.4014 0.03194 0.3194 0.0005831      0.0005657 1.0009
-## sigeta 0.6722 0.02146 0.2146 0.0003919      0.0003691 1.0005
+## a      0.1184 0.01726 0.1726 0.0003151      0.0003410 1.0021
+## b      0.4858 0.02774 0.2774 0.0005065      0.0005065 1.0019
+## k      4.9005 0.03196 0.3196 0.0005835      0.0005836 0.9997
+## psi    1.4030 0.03158 0.3158 0.0005765      0.0005666 1.0003
+## sigeta 0.6720 0.02096 0.2096 0.0003826      0.0004017 1.0027
 ## 
 ## 2. Quantiles for each variable:
 ## 
 ##           2.5%    25%    50%    75%  97.5%
-## a      0.08734 0.1059 0.1175 0.1293 0.1538
-## b      0.43155 0.4686 0.4870 0.5065 0.5429
-## k      4.83922 4.8784 4.9000 4.9203 4.9608
-## psi    1.33744 1.3791 1.4019 1.4234 1.4607
-## sigeta 0.63069 0.6572 0.6719 0.6868 0.7147
+## a      0.08774 0.1066 0.1176 0.1295 0.1539
+## b      0.43299 0.4670 0.4854 0.5042 0.5401
+## k      4.84050 4.8774 4.9002 4.9230 4.9622
+## psi    1.34053 1.3818 1.4038 1.4234 1.4657
+## sigeta 0.63229 0.6578 0.6719 0.6859 0.7142
 ```
 
 Let's represent survival as a function of age at the population level. 
[…]
