### 2018-01-07 217fff8
Delete paper.bib
