+++ paper/paper.bib
@@ -1,57 +1,0 @@
-% Encoding: UTF-8
-
-@Manual{R,
-  title        = {R: A Language and Environment for Statistical Computing},
-  author       = {{R Core Team}},
-  organization = {R Foundation for Statistical Computing},
-  address      = {Vienna, Austria},
-  year         = {2016},
-  url          = {http://www.R-project.org/},
-}
-
-@Article{LeleDennisLutscher2007,
-  author        = {Lele, SR and Dennis, D and Lutscher, F},
-  title         = {Data cloning: {E}asy maximum likelihood estimation for complex ecological models using {B}ayesian {M}arkov chain {M}onte {C}arlo methods},
-  journal       = {Ecology Letters},
-  year          = {2007},
-  volume        = {10},
-  pages         = {551–563},
-  __markedentry = {[bluteauk:1]},
-  doi           = {10.1111/j.1461-0248.2007.01047.x},
-}
-
-@Article{LeleNadeemSchmuland2010,
-  author        = {Lele, SR and Nadeem, K and Schmuland, B},
-  title         = {Estimability and Likelihood Inference for {G}eneralized {L}inear {M}ixed {M}odels Using {D}ata {C}loning},
-  journal       = {Journal of the American Statistical association},
-  year          = {2010},
-  volume        = {105},
-  pages         = {1617–-1625},
-  __markedentry = {[bluteauk:1]},
-  doi           = {10.1198/jasa.2010.tm09757},
-}
-
-@Article{MarzolinCharmantierGimenez2011,
-  author        = {Marzolin, G and Charmantier, A and Gimenez, O},
-  title         = {Frailty in state-space models: {A}pplication to actuarial senescence in the {D}ipper},
-  journal       = {Ecology},
-  year          = {2011},
-  volume        = {92},
-  pages         = {562–-567},
-  __markedentry = {[bluteauk:1]},
-  doi           = {10.1890/10-0306.1},
-}
-
-@Article{Missov2013,
-  author        = {Missov, TI},
-  title         = {Gamma-{G}ompertz life expectancy at birth},
-  journal       = {Demographic Research},
-  year          = {2013},
-  volume        = {28},
-  pages         = {259–-270},
-  __markedentry = {[bluteauk:1]},
-  doi           = {10.4054/DemRes.2013.28.9},
-}
-
-
-@Comment{jabref-meta: databaseType:bibtex;}
