### 2018-01-07 3ff360e
Delete paper.md
