+++ README.md
@@ -7,6 +7,9 @@ output:
     keep_md: yes
 ---
 
+
+
+
 # In brief
   
 `GammaGompertzCR` is an R package that allows estimating survival in free-ranging animal populations using a Gompertz capture-recapture model with a Gamma frailty to deal with individual heterogeneity. It uses data cloning in the Bayesian framework to get maximum likelihood parameter estimates (Lele et al. 2007). Data cloning uses multiple copies of the data to produce prior-invariant inferences and a normal distribution centered at the maximum likelihood estimates. In addition, this method allows detecting non-identifiable parameters (Lele et al. 2010). To use the package, users should be familiar with Bayesian MCMC techniques and in particular how to interpret convergence diagnostics. We refer to Robert and Casella (2010) for an introduction to Bayesian MCMC techniques with R and to King et al. (2009) for an introduction for ecologists.
@@ -105,20 +108,20 @@ summary(post_inf[[length(clo)]])
 ##    plus standard error of the mean:
 ## 
 ##          Mean      SD  DC SD  Naive SE Time-series SE  R hat
-## a      0.1184 0.01726 0.1726 0.0003151      0.0003410 1.0021
-## b      0.4858 0.02774 0.2774 0.0005065      0.0005065 1.0019
-## k      4.9005 0.03196 0.3196 0.0005835      0.0005836 0.9997
-## psi    1.4030 0.03158 0.3158 0.0005765      0.0005666 1.0003
-## sigeta 0.6720 0.02096 0.2096 0.0003826      0.0004017 1.0027
+## a      0.1185 0.01704 0.1704 0.0003111      0.0003163 1.0016
+## b      0.4858 0.02880 0.2880 0.0005258      0.0005803 0.9996
+## k      4.8993 0.03116 0.3116 0.0005688      0.0005794 1.0000
+## psi    1.4007 0.03173 0.3173 0.0005793      0.0005223 1.0028
+## sigeta 0.6725 0.02079 0.2079 0.0003795      0.0003693 1.0018
 ## 
 ## 2. Quantiles for each variable:
 ## 
 ##           2.5%    25%    50%    75%  97.5%
-## a      0.08774 0.1066 0.1176 0.1295 0.1539
-## b      0.43299 0.4670 0.4854 0.5042 0.5401
-## k      4.84050 4.8774 4.9002 4.9230 4.9622
-## psi    1.34053 1.3818 1.4038 1.4234 1.4657
-## sigeta 0.63229 0.6578 0.6719 0.6859 0.7142
+## a      0.08709 0.1066 0.1178 0.1299 0.1530
+## b      0.43066 0.4668 0.4853 0.5049 0.5445
+## k      4.83784 4.8783 4.8990 4.9193 4.9624
+## psi    1.33890 1.3797 1.4004 1.4227 1.4615
+## sigeta 0.63368 0.6578 0.6717 0.6869 0.7138
 ```
 
 Let's represent survival as a function of age at the population level. 
@@ -134,7 +137,7 @@ S = exp(-k*log(1+(exp(b*grid_age)-1)*a/(b*k)))
 plot(grid_age,S,xlab='age',ylab='estimated survival',lwd=2,type='l')
 ```
 
-![](README_files/figure-html/unnamed-chunk-6-1.png)<!-- -->
+![](README_files/figure-html/unnamed-chunk-7-1.png)<!-- -->
 
 ## Convergence diagnostics
 
@@ -144,41 +147,41 @@ First have a look at the traceplots for all parameter estimates got with the las
[…]
