+++ paper/paper.bib
@@ -1,0 +1,69 @@
+@book{king2009bayesian,
+  title={Bayesian Analysis for Population Ecology},
+  author={King, R. and Gimenez, O. and Morgan, B. and Brooks, S.},
+  year={2009},
+  publisher={Chapman \& Hall/CRC}
+}
+
+@book{robert2010introducing,
+  title={Introducing Monte Carlo Methods with R},
+  author={Robert, C. and Casella, G.},
+  series={Use R!},
+  year={2010},
+  publisher={Springer}
+}
+
+@Manual{R,
+  title        = {R: A Language and Environment for Statistical Computing},
+  author       = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
+  address      = {Vienna, Austria},
+  year         = {2016},
+  url          = {http://www.R-project.org/}
+}
+
+@Article{LeleDennisLutscher2007,
+  author        = {Lele, SR and Dennis, D and Lutscher, F},
+  title         = {Data cloning: {E}asy maximum likelihood estimation for complex ecological models using {B}ayesian {M}arkov chain {M}onte {C}arlo methods},
+  journal       = {Ecology Letters},
+  year          = {2007},
+  volume        = {10},
+  pages         = {551-563}
+}
+
+@Article{LeleNadeemSchmuland2010,
+  author        = {Lele, SR and Nadeem, K and Schmuland, B},
+  title         = {Estimability and Likelihood Inference for {G}eneralized {L}inear {M}ixed {M}odels Using {D}ata {C}loning},
+  journal       = {Journal of the American Statistical association},
+  year          = {2010},
+  volume        = {105},
+  pages         = {1617-1625}
+}
+
+@Article{MarzolinCharmantierGimenez2011,
+  author        = {Marzolin, G and Charmantier, A and Gimenez, O},
+  title         = {Frailty in state-space models: {A}pplication to actuarial senescence in the {D}ipper},
+  journal       = {Ecology},
+  year          = {2011},
+  volume        = {92},
+  pages         = {562-567}
+}
+
+@Article{Missov2013,
+  author        = {Missov, TI},
+  title         = {Gamma-{G}ompertz life expectancy at birth},
+  journal       = {Demographic Research},
+  year          = {2013},
+  volume        = {28},
+  pages         = {259-270}
+}
+
+@Article{Solymos2010,
+   author        = {Solymos, P},
+   title         = {dclone: Data Cloning in R},
+   journal       = {The R Journal},
+   year          = {2010},
+   volume        = {2},
+   pages         = {29-37}
+}
+
