+++ R/zzz.R
@@ -1,0 +1,6 @@
+.onAttach <-
+  function(libname = find.package("GammaGompertzCR"),
+           pkgname = "GammaGompertzCR") {
+    packageStartupMessage("Thanks for using package GammaGompertzCR. 
+    Check out https://github.com/oliviergimenez/GammaGompertzCR for an example.")
+}
