### 2018-01-08 #216 @oliviergimenez
Dear @bbolker and @karthik,
First, apologies for taking so long to get back to you. The second semester of 2017 has been hectic. At last we had the chance to go through @bbolker's review over the Christmas break. @bbolker, our answers to your comments are attached. We have also submitted a new version of the package.
Thanks a lot for your time in reviewing our work.
Olivier & Gilbert

[bolker_review_with_answers.pdf](https://github.com/openjournals/joss-reviews/files/1612177/bolker_review_with_answers.pdf)
