### 2018-01-08 #216 @bbolker
These responses seem fine to me. I will go through and close the issues.
