### 2018-01-08 oliviergimenez/GammaGompertzCR#4 @bbolker
The authors say:

> The package contains only 3 functions and is basically a wrapper for the dcmle::jags.parfit function for fitting a Gompertz model to capture-recapture data. We now provide in the Readme and paper the results of the full analyses of a dataset (numerical outputs and figures) that can be used as a validation for users. Hopefully, this will do the same job as unit tests would do.

I still don't think this really fulfils the task of tests (either unit or integration). How will you know in future if changes to JAGS, or changes that you make to the code, keep the behaviour consistent?
