### 2018-01-08 oliviergimenez/GammaGompertzCR#5 @bbolker
The README example doesn't work without running `library(dclone)` in the session first.

How is the number of cores for a parallel run determined? Is this something you want to give some hints to the user about? (On my system, without setting any options about the number of cores available, the example takes 8 minutes to run ...)

The style in the README file isn't completely consistent; for examples, assignments are sometimes `=` and sometimes `<-` (I prefer ` <- ` , i.e. arrows with spaces around them ...)

I *think* the code on l. 208 of the README file is intended to spit out a string that can be cut-and-pasted into the R console? (1) `paste(sprintf("post_inf[[%d]]",seq_along(clo)),collapse="")` would do the same thing slightly more elegantly ...); (2) `do.call(dctable,post_inf)` is the simple, idiomatic way to do this  ... or convince the `dclone` maintainers to allow `x` to be a list, which would then be expanded ...
