### 2018-01-08 oliviergimenez/GammaGompertzCR#6 @bbolker
I'm satisfied with the authors' responses here.
