### 2018-01-08 #216 @bbolker
I've left a couple of specific comments on the issues list. Everything else was fine.
