### 2018-01-08 oliviergimenez/GammaGompertzCR#4 @oliviergimenez
By comparing the new results to the results provided in the paper (Table 1) I guess. It would mean making sure that the data haven't changed, by including a test maybe? Apologies if my answer sounds naive, I'm not so familiar with testing.
