### 2018-01-08 oliviergimenez/GammaGompertzCR#4 @bbolker
I guess I'm looking for something more automated: you say `./run_test` from the command line and it runs some code and says "yes, these results match the reference results I have stored" or "no, tests failed". It would fail if the data changed *or* the algorithm changed. You could always update the stored reference results if you made a deliberate change in the data or in the algorithm.
