### 2018-01-08 oliviergimenez/GammaGompertzCR#4 @oliviergimenez
I understand. I'll work on it. Thanks!
