+++ README.md
@@ -7,9 +7,6 @@ output:
     keep_md: yes
 ---
 
-
-
-
 # In brief
   
 `GammaGompertzCR` is an R package that allows estimating survival in free-ranging animal populations using a Gompertz capture-recapture model with a Gamma frailty to deal with individual heterogeneity. It uses data cloning in the Bayesian framework to get maximum likelihood parameter estimates (Lele et al. 2007). Data cloning uses multiple copies of the data to produce prior-invariant inferences and a normal distribution centered at the maximum likelihood estimates. In addition, this method allows detecting non-identifiable parameters (Lele et al. 2010). To use the package, users should be familiar with Bayesian MCMC techniques and in particular how to interpret convergence diagnostics. We refer to Robert and Casella (2010) for an introduction to Bayesian MCMC techniques with R and to King et al. (2009) for an introduction for ecologists.
@@ -69,16 +66,16 @@ head(dipper_tiny)
 Now let's fit the Gamma-Gompertz model to these capture-recapture data:
 
 ```r
-clo = c(1,10,50,100) # numbers of clones
-nu = 1000 # number of updates
-ni = 5000 # number of iterations
-nt = 5 # thinning
-nc = 3 # number of chains
-initmeans = c(0.15,0.5,4.9,1.4,-0.4) # means of normally distributed priors for parameters
-initprec = 1000 # common precision of the normal priors for the first number of clones
+clo <- c(1,10,50,100) # numbers of clones
+nu <- 1000 # number of updates
+ni <- 5000 # number of iterations
+nt <- 5 # thinning
+nc <- 3 # number of chains
+initmeans <- c(0.15,0.5,4.9,1.4,-0.4) # means of normally distributed priors for parameters
+initprec <- 1000 # common precision of the normal priors for the first number of clones
 
 # fitting procedure (this may take a few minutes)
-post_inf = fit_ggcr(dipper_tiny,clo,nu,ni,nt,nc,initmeans,initprec)
+post_inf <- fit_ggcr(dipper_tiny,clo,nu,ni,nt,nc,initmeans,initprec)
 ```
 
 ```
@@ -113,20 +110,20 @@ summary(post_inf[[length(clo)]])
 ##    plus standard error of the mean:
 ## 
 ##          Mean      SD  DC SD  Naive SE Time-series SE  R hat
-## a      0.1185 0.01723 0.1723 0.0003145      0.0003225 0.9995
-## b      0.4868 0.02826 0.2826 0.0005160      0.0005081 1.0010
-## k      4.8995 0.03179 0.3179 0.0005805      0.0005688 1.0013
-## psi    1.4014 0.03241 0.3241 0.0005917      0.0005918 0.9996
-## sigeta 0.6724 0.02091 0.2091 0.0003818      0.0003763 1.0022
+## a      0.1182 0.01689 0.1689 0.0003083      0.0003271 1.0047
+## b      0.4861 0.02858 0.2858 0.0005219      0.0005378 1.0014
+## k      4.8997 0.03112 0.3112 0.0005682      0.0005838 1.0005
+## psi    1.4025 0.03171 0.3171 0.0005790      0.0005900 1.0002
+## sigeta 0.6721 0.02067 0.2067 0.0003774      0.0003914 0.9996
 ## 
 ## 2. Quantiles for each variable:
 ## 
 ##           2.5%    25%    50%    75%  97.5%
-## a      0.08706 0.1067 0.1174 0.1297 0.1542
[…]
