+++ NAMESPACE
@@ -3,6 +3,7 @@
 export(fit_ggcr)
 export(ggcr_model)
 export(prep_data)
+import(dclone)
 import(dcmle)
 import(parallel)
 importFrom(stats,sd)
