+++ README.md
@@ -110,20 +110,20 @@ summary(post_inf[[length(clo)]])
 ##    plus standard error of the mean:
 ## 
 ##          Mean      SD  DC SD  Naive SE Time-series SE  R hat
-## a      0.1182 0.01689 0.1689 0.0003083      0.0003271 1.0047
-## b      0.4861 0.02858 0.2858 0.0005219      0.0005378 1.0014
-## k      4.8997 0.03112 0.3112 0.0005682      0.0005838 1.0005
-## psi    1.4025 0.03171 0.3171 0.0005790      0.0005900 1.0002
-## sigeta 0.6721 0.02067 0.2067 0.0003774      0.0003914 0.9996
+## a      0.1187 0.01721 0.1721 0.0003142      0.0003184 0.9996
+## b      0.4858 0.02840 0.2840 0.0005184      0.0005186 0.9998
+## k      4.8999 0.03184 0.3184 0.0005814      0.0005729 1.0010
+## psi    1.4031 0.03177 0.3177 0.0005801      0.0005802 1.0020
+## sigeta 0.6723 0.02120 0.2120 0.0003871      0.0003871 0.9999
 ## 
 ## 2. Quantiles for each variable:
 ## 
 ##           2.5%    25%    50%    75%  97.5%
-## a      0.08783 0.1062 0.1174 0.1291 0.1538
-## b      0.43088 0.4673 0.4862 0.5055 0.5430
-## k      4.83993 4.8797 4.8995 4.9201 4.9617
-## psi    1.34211 1.3806 1.4019 1.4239 1.4660
-## sigeta 0.63315 0.6579 0.6717 0.6861 0.7140
+## a      0.08708 0.1068 0.1181 0.1298 0.1546
+## b      0.43072 0.4671 0.4856 0.5046 0.5423
+## k      4.83788 4.8780 4.9004 4.9218 4.9615
+## psi    1.34099 1.3821 1.4027 1.4248 1.4677
+## sigeta 0.63147 0.6576 0.6718 0.6864 0.7150
 ```
 
 Let's represent survival as a function of age at the population level. 
@@ -140,6 +140,10 @@ plot(grid_age,S,xlab='age',ylab='estimated survival',lwd=2,type='l')
 ```
 
 ![](README_files/figure-html/unnamed-chunk-6-1.png)<!-- -->
+
+## How to choose the precision, number of iterations and clones?
+
[…]
