### 2018-01-10 oliviergimenez/GammaGompertzCR#5 @oliviergimenez
Hi @bbolker,
The README file is working now.
We have added a section `How to choose the precision, number of iterations and clones?` to hints to the user.
The style was fixed, ` <- ` is used everywhere.
Thanks for your tip on using `do.call`.
Now working on the tests.
Thanks for your help.
