### 2018-01-12 oliviergimenez/GammaGompertzCR#4 @oliviergimenez
Hi @bbolker,
I have added a test for the main function fit_ggcr; devtools::test() runs ok.
Cheers,
Olivier
