### 2018-01-12 oliviergimenez/GammaGompertzCR#4 @bbolker
great!
