### 2018-01-12 oliviergimenez/GammaGompertzCR#5 @bbolker
sounds good.
