### 2018-01-12 #216 @oliviergimenez
Dear @bbolker, cc @karthik,
Thanks for closing all issues, your help has been very valuable.
Olivier & Gilbert
