### 2018-01-18 #216 @karthik
Hi @oliviergimenez
Congrats on getting your paper accepted! 🎉
At this point I would like you to archive your software in Zenodo and post the DOI here. Then I'll be able to set the archive and complete the final steps.
