+++ README.md
@@ -1,17 +1,11 @@
----
-title: "Fitting a Gamma-Gompertz survival model to capture-recapture data collected on free-ranging animal populations"
-author: "Gilbert Marzolin and Olivier Gimenez"
-date: "12 January 2018"
-output: 
-  html_document: 
-    keep_md: yes
----
+# Fitting a Gamma-Gompertz survival model to capture-recapture data collected on free-ranging animal populations
+Gilbert Marzolin and Olivier Gimenez
 
-# In brief
+## In brief
   
 `GammaGompertzCR` is an R package that allows estimating survival in free-ranging animal populations using a Gompertz capture-recapture model with a Gamma frailty to deal with individual heterogeneity. It uses data cloning in the Bayesian framework to get maximum likelihood parameter estimates (Lele et al. 2007). Data cloning uses multiple copies of the data to produce prior-invariant inferences and a normal distribution centered at the maximum likelihood estimates. In addition, this method allows detecting non-identifiable parameters (Lele et al. 2010). To use the package, users should be familiar with Bayesian MCMC techniques and in particular how to interpret convergence diagnostics. We refer to Robert and Casella (2010) for an introduction to Bayesian MCMC techniques with R and to King et al. (2009) for an introduction for ecologists.
 
-# Installation
+## Installation
 
 This repository hosts the development version of the package. Assuming `JAGS` is already on your system (otherwise it'll take a minute to [get it](http://mcmc-jags.sourceforge.net/), you can install `GammaGompertzCR` as follows:
 
@@ -21,7 +15,7 @@ library("devtools")
 install_github('oliviergimenez/GammaGompertzCR')
 ```
 
-# Getting help, reporting an issue or contribute
+## Getting help, reporting an issue or contribute
 
 To report bugs/issues/feature requests, please file an [issue](https://github.com/oliviergimenez/GammaGompertzCR/issues). These are very welcome!
 
@@ -29,9 +23,9 @@ Feel free to contribute to the repository by forking and submitting a pull reque
 
 If you prefer to email, feel free to drop a message to the Gilbert and Olivier (see the DESCRIPTION file of this repository for their contact details).
 
-# A session example
+## A session example
 
-## Preliminary steps
+### Preliminary steps
 
 First, we need to load the package:
 
@@ -61,7 +55,7 @@ head(dipper_tiny)
 ## [6,]    1    1    1    2    2    2    2    2    2
 ```
 
-## Model fitting
+### Model fitting
 
 Now let's fit the Gamma-Gompertz model to these capture-recapture data:
 
@@ -141,11 +135,11 @@ plot(grid_age,S,xlab='age',ylab='estimated survival',lwd=2,type='l')
 
 ![](README_files/figure-html/unnamed-chunk-6-1.png)<!-- -->
 
-## How to choose the precision, number of iterations and clones?
+### How to choose the precision, number of iterations and clones?
 
[…]
